--- emptyWindow/emptyWindow.cpp
@@ -1,33 +1,0 @@
-#include <QtGui>
-#include <QtOpenGL>
-#include<glut.h>
-#include "emptyWindow.h"
-
-EmptyWindow::EmptyWindow(QWidget *parent)
-    : QGLWidget(parent)
-{
-    setFormat(QGLFormat(QGL::DoubleBuffer | QGL::DepthBuffer));
-}
-
-void EmptyWindow::initializeGL()
-{
-    qglClearColor(Qt::blue);
-    glShadeModel(GL_FLAT);
-    glEnable(GL_DEPTH_TEST);
-    glEnable(GL_CULL_FACE);
-}
-
-void EmptyWindow::resizeGL(int width, int height)
-{
-    glViewport(0, 0, width, height);
-    glMatrixMode(GL_PROJECTION);
-    glLoadIdentity();
-    GLfloat x = GLfloat(width) / height;
-    glFrustum(-x, +x, -1.0, +1.0, 4.0, 15.0);
-    glMatrixMode(GL_MODELVIEW);
-}
-
-void EmptyWindow::paintGL()
-{
-    glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-}
